refactor: move design-tokens to utils
