+++ src/components/index.ts
@@ -1,7 +1,6 @@
 export * from './aspect-ratio';
 export * from './button';
 export * from './container';
-export * from './design-tokens';
 export * from './image';
 export * from './layout';
 export * from './link';
