+++ src/utilities/index.ts
@@ -1,1 +1,2 @@
 export { PropsWithIdAndClassname } from './common';
+export * from './design-tokens';
